+++ circuits/src/verifier/accumulator.rs
@@ -242,6 +242,10 @@ impl<S: SelfEmulation> Instantiable<S::F> for AssignedAccumulator<S> {
         .flatten()
         .collect()
     }
+
+    fn from_public_input(_fields: &[S::F]) -> Option<Accumulator<S>> {
+        unimplemented!("Size of inner MSMs cannot be known from public input format.")
+    }
 }
 
 impl<S: SelfEmulation> AssignedAccumulator<S> {